Refactors edit file metadata to use modal
Updates the edit file functionality to utilize a Bootstrap modal for a better user experience.

Replaces the direct edit page with a modal that is populated via JavaScript, enhancing responsiveness and providing a cleaner interface.

Improves data handling by using FormData for submission and JSON responses for feedback.

Adds new CSS to style the upload and edit file modals.